🗺️ Atlas: Extract AST depth validation into `semantic/validation.rs`
